CM-1632: Implemented LoggedArtifactAssetImpl. Fixed DataModelUtils.metadataFromJson to use explicit type reference.

+++ comet-java-client/src/main/java/ml/comet/experiment/artifact/LoggedArtifactAsset.java
@@ -48,7 +48,7 @@ public interface LoggedArtifactAsset {
      *
      * @return the optional size of the asset file.
      */
-    Optional<Integer> getFileSize();
+    Optional<Long> getFileSize();
 
     /**
      * Returns the optional {@link URI} of the asset if this is remote asset.
